dotty fixes
