+++ izumi-reflect/izumi-reflect/src/test/scala-2/izumi/reflect/test/LightTypeTagTest.scala
@@ -26,6 +26,25 @@ class LightTypeTagTest extends SharedLightTypeTagTest {
 
   "lightweight type tags (Scala 2)" should {
 
+    "wildcards are not supported (wildcard=Any, historically due to behavior of .dealias on 2.12/13, see scala.reflect.internal.tpe.TypeMaps#ExistentialExtrapolation)" in {
+      assertDifferent(LTT[Set[_]], LTT[Set[Any]])
+      assertDifferent(LTT[List[_]], LTT[List[Any]])
+
+      assertChild(LTT[Set[Int]], LTT[Set[_]])
+      assertNotChild(LTT[Set[_]], LTT[Set[Int]])
+
+      assertChild(LTT[List[Int]], LTT[List[_]])
+      assertNotChild(LTT[List[_]], LTT[List[Int]])
+
+      assertChild(LTT[Int => Int], LTT[_ => Int])
+    }
+
+    "wildcards with bounds are not supported (upper bound is the type, historically due to behavior of .dealias on 2.12/13, see scala.reflect.internal.tpe.TypeMaps#ExistentialExtrapolation)" in {
+      assertDifferent(LTT[Option[W1]], LTT[Option[_ <: W1]])
+      assertDifferent(LTT[Option[H2]], LTT[Option[_ >: H4 <: H2]])
+      assertDifferent(LTT[Option[Any]], LTT[Option[_ >: H4]])
+    }
+
     "support typetag combination (Scala 2 syntax)" in {
       assertCombine(`LTT[_[_]]`[T0[Id, *[_]]], `LTT[_]`[FP], LTT[T0[Id, FP]])
     }
